# Conflicts:
#	features/top_ads/top_ads_management_common/src/main/java/com/tokopedia/topads/common/analytics/TopAdsCreateAnalytics.kt

+++ features/fintech/paylater/src/main/java/com/tokopedia/pdpsimulation/paylater/viewModel/PayLaterViewModel.kt
@@ -33,7 +33,7 @@ class PayLaterViewModel @Inject constructor(
 
     private var idlingResourceProvider = TkpdIdlingResourceProvider.provideIdlingResource("SIMULATION")
 
-    fun getPayLaterSimulationData(amount: Int) {
+    fun getPayLaterSimulationData(amount: Long) {
         idlingResourceProvider?.increment()
         payLaterSimulationDataUseCase.cancelJobs()
         payLaterSimulationDataUseCase.getSimulationData(
